FEAT: 관리자 판매관리 기능 완료, 필터링 페이징 기능 보완
관리자 맨패관리 부분, 상품등록, 상품삭제/수정, 상품문의 기능 완료, 필터링 하는 관리자 게시글 부분 페이징 기능 추가

+++ JMC/src/main/java/com/kh/jaManChw/dto/ItemOption.java
@@ -4,13 +4,17 @@ public class ItemOption {
 
 	private int itemOptionno;
 	private int itemno;
+	
+	//Option이 배열값을 받는 경우가 있기에 - 멤버 필드를 배열로 받고자함
+	//뭔가 문제점이 많음...그냥 Controller에서 받을때 배열로 만들고 
+	// controller에서 나눠서 리스트로된 DTO를 만블어보자..!
 	private String optionContent;
 	private int optionCount;
-	private double extraCharge;
+	private int extraCharge;
 	
 	public ItemOption() {}
 
-	public ItemOption(int itemOptionno, int itemno, String optionContent, int optionCount, double extraCharge) {
+	public ItemOption(int itemOptionno, int itemno, String optionContent, int optionCount, int extraCharge) {
 		super();
 		this.itemOptionno = itemOptionno;
 		this.itemno = itemno;
@@ -57,11 +61,11 @@ public class ItemOption {
 		this.optionCount = optionCount;
 	}
 
-	public double getExtraCharge() {
+	public int getExtraCharge() {
 		return extraCharge;
 	}
 
-	public void setExtraCharge(double extraCharge) {
+	public void setExtraCharge(int extraCharge) {
 		this.extraCharge = extraCharge;
 	}
